1. 安裝npm; 2. 安裝 Laravel 的 ui 工具套件： composer require laravel/ui --dev 3. 建立含有使用者認證的bootstrap框架 php artisan ui bootstrap --auth 4. 安裝相依性套件並編譯 npm install && npm run dev npm run development 5. 編輯 app/User.php，加入三種帳號角色名稱的常數字串，並將 role 這個欄位加入 fillable 中。 6. 編輯 database/migrations/2014_10_12_000000_create_users_table.php 這一個建立使用者資料表的 migration 設定檔，在其中加入一欄用於儲存帳號角色的 role 欄位。 7. 編輯 app/Http/Controllers/Auth/RegisterController.php，設定使用者註冊時的角色都是一般使用者。 8. 建立 Gate 權限規則。 9.自動生成三個 users: admin, manager, and user (方便我們做測試之用) 10.更新兩個views： roles.index 和areas.index，讓他們針對admin 與 manager 呈現受控管的超連結(新增、修改、刪除) 11.讓每一個路由都必須經過 auth 的 middleware (即必須登入帳號才可以使用)。 12.增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用 13.不允許編輯表單給 user 權限使用，但允許 admin或manager使用

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AreasController;
 use App\Http\Controllers\RolesController;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -13,6 +14,7 @@ use App\Http\Controllers\RolesController;
 | contains the "web" middleware group. Now create something great!
 |
 */
+Route::middleware(['auth'])->group(function () {
 Route::get("/",function(){
     return redirect('roles');
 });
@@ -32,21 +34,21 @@ Route::get('roles/{id}', [RolesController::class, 'show'])->where('id', '[0-9]+'
 // 修改單一角色表單
 Route::get('roles/{id}/edit', [RolesController::class, 'edit'])->where('id', '[0-9]+')->name('roles.edit');
 // 刪除單一球員資料
-Route::delete('roles/delete/{id}', [RolesController::class, 'destroy'])->where('id', '[0-9]+')->name('roles.destroy');
+Route::delete('roles/delete/{id}', [RolesController::class, 'destroy'])->where('id', '[0-9]+')->name('roles.destroy')->middleware('can:admin');
 // 刪除單一球隊及旗下球員資料
-Route::delete('areas/delete/{id}', [AreasController::class, 'destroy'])->where('id', '[0-9]+')->name('areas.destroy');
+Route::delete('areas/delete/{id}', [AreasController::class, 'destroy'])->where('id', '[0-9]+')->name('areas.destroy')->middleware('can:admin');
 // 新增角色表單
-Route::get('roles/create', [RolesController::class, 'create'])->name('roles.create');
+Route::get('roles/create',[RolesController::class, 'create'])->name('roles.create')->middleware('can:admin');
 // 新增區域表單
-Route::get('areas/create', [AreasController::class, 'create'])->name('areas.create');
+Route::get('areas/create',[AreasController::class, 'create'])->name('areas.create')->middleware('can:admin');
 // 修改角色資料
 Route::patch('roles/update/{id}', [RolesController::class, 'update'])->where('id', '[0-9]+')->name('roles.update');
 // 修改區域資料
 Route::patch('areas/update/{id}', [AreasController::class, 'update'])->where('id', '[0-9]+')->name('areas.update');
 // 儲存新角色資料
-Route::post('roles/store', [RolesController::class, 'store'])->where('id', '[0-9]+')->name('roles.store');
+Route::post('roles/store', [RolesController::class, 'store'])->where('id', '[0-9]+')->name('roles.store')->middleware('can:admin');
 // 儲存新區域資料
-Route::post('areas/store', [AreasController::class, 'store'])->name('areas.store');
+Route::post('areas/store', [AreasController::class, 'store'])->where('id', '[0-9]+')->name('areas.store')->middleware('can:admin');
 
 Route::get('areas/land',[AreasController::class,'land'])->name('areas.land');
 Route::get('areas/sea',[AreasController::class,'sea'])->name('areas.sea');
@@ -55,3 +57,8 @@ Route::get('roles/fourstars',[RolesController::class,'fourstars'])->name('roles.
 Route::get('roles/fivestars',[RolesController::class,'fivestars'])->name('roles.fivestars');
 
 Route::get('roles/gender', [RolesController::class, 'gender'])->name('roles.gender');
+
+});
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
